+++ src/autocomplete/provider.test.ts
@@ -1,55 +1,55 @@
 import { strict as assert } from "assert"
 
-import * as input from "./input"
+import * as provider from "./provider"
 
 suite("analyzeInput", function() {
   suite("markdown link", function() {
     test("middle of sentence", function() {
-      const have = input.analyze(
+      const have = provider.determineType(
         "Check out [the new and here is another [one](one.md)",
         18
       )
-      const want = input.AutocompleteType.MD_LINK
+      const want = provider.AutocompleteType.MD_LINK
       assert.deepEqual(have, want)
     })
 
     test("starting", function() {
-      const have = input.analyze("[", 1)
-      const want = input.AutocompleteType.MD_LINK
+      const have = provider.determineType("[", 1)
+      const want = provider.AutocompleteType.MD_LINK
       assert.deepEqual(have, want)
     })
   })
 
   suite("image tag", function() {
     test("middle of sentence", function() {
-      const have = input.analyze(
+      const have = provider.determineType(
         "Check out ![the and here is another [one](one.md)",
         15
       )
-      const want = input.AutocompleteType.IMG
+      const want = provider.AutocompleteType.IMG
       assert.deepEqual(have, want)
     })
 
     test("starting", function() {
-      const have = input.analyze("![", 2)
-      const want = input.AutocompleteType.IMG
+      const have = provider.determineType("![", 2)
+      const want = provider.AutocompleteType.IMG
       assert.deepEqual(have, want)
     })
   })
 
   suite("heading", function() {
     test("middle of sentence", function() {
-      const have = input.analyze(
+      const have = provider.determineType(
         "Check out #the and here is another [one](one.md)",
         14
       )
-      const want = input.AutocompleteType.NONE
+      const want = provider.AutocompleteType.NONE
       assert.deepEqual(have, want)
     })
 
     test("starting", function() {
-      const have = input.analyze("#", 1)
-      const want = input.AutocompleteType.HEADING
+      const have = provider.determineType("#", 1)
+      const want = provider.AutocompleteType.HEADING
       assert.deepEqual(have, want)
     })
   })
